+++ interfaces/yearn/IEarnableVault.sol
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.6.12;
+
 interface IEarnableVault {
     function earn() external;
-    function available() external view returns (uint _available);
+
+    function available() external view returns (uint256 _available);
 }
